refactor: move setupButton and setupInteractions to dedicated packages
