+++ packages/ng-primitives/button/src/index.ts
@@ -1,2 +1,3 @@
 export { NgpButton } from './button/button';
 export { injectButtonState, provideButtonState } from './button/button-state';
+export { setupButton } from './button/setup-button';
