+++ packages/ng-primitives/combobox/src/combobox-button/combobox-button.ts
@@ -1,5 +1,6 @@
 import { computed, Directive, HostListener, input } from '@angular/core';
-import { injectElementRef, setupInteractions } from 'ng-primitives/internal';
+import { setupInteractions } from 'ng-primitives/interactions';
+import { injectElementRef } from 'ng-primitives/internal';
 import { uniqueId } from 'ng-primitives/utils';
 import { injectComboboxState } from '../combobox/combobox-state';
