+++ packages/ng-primitives/combobox/src/combobox-input/combobox-input.ts
@@ -1,6 +1,7 @@
 import { computed, Directive, HostListener, input } from '@angular/core';
 import { setupFormControl } from 'ng-primitives/form-field';
-import { injectElementRef, setupInteractions } from 'ng-primitives/internal';
+import { setupInteractions } from 'ng-primitives/interactions';
+import { injectElementRef } from 'ng-primitives/internal';
 import { uniqueId } from 'ng-primitives/utils';
 import { injectComboboxState } from '../combobox/combobox-state';
