+++ packages/ng-primitives/date-picker/src/date-picker-date-button/date-picker-date-button.ts
@@ -1,7 +1,7 @@
 import { FocusMonitor } from '@angular/cdk/a11y';
 import { computed, Directive, ElementRef, HostListener, inject, OnDestroy } from '@angular/core';
+import { setupButton } from 'ng-primitives/button';
 import { injectDateAdapter } from 'ng-primitives/date-time';
-import { setupButton } from 'ng-primitives/internal';
 import { injectDatePickerCellDate } from '../date-picker-cell-render/date-picker-cell-render-token';
 import { injectDateControllerState } from '../date-picker/date-picker-state';
 import { NgpDatePickerDateButtonToken } from './date-picker-date-button-token';
