+++ packages/ng-primitives/input/src/input/input.ts
@@ -2,7 +2,8 @@ import { BooleanInput } from '@angular/cdk/coercion';
 import { booleanAttribute, Directive, input, Signal } from '@angular/core';
 import { NgpAutofill } from 'ng-primitives/autofill';
 import { setupFormControl } from 'ng-primitives/form-field';
-import { injectElementRef, setupInteractions } from 'ng-primitives/internal';
+import { setupInteractions } from 'ng-primitives/interactions';
+import { injectElementRef } from 'ng-primitives/internal';
 import { injectSearchState } from 'ng-primitives/search';
 import { NgpControlStatus, uniqueId } from 'ng-primitives/utils';
 import { inputState, provideInputState } from './input-state';
