+++ packages/ng-primitives/interactions/src/index.ts
@@ -3,3 +3,4 @@ export { NgpFocus } from './focus/focus';
 export { NgpHover } from './hover/hover';
 export { NgpMove, NgpMoveEvent } from './move/move';
 export { NgpPress } from './press/press';
+export { setupInteractions } from './setup-interactions/setup-interactions';
