+++ packages/ng-primitives/interactions/src/setup-interactions/interaction.ts
