+++ packages/ng-primitives/internal/src/interactions/index.ts
@@ -1,5 +1,4 @@
 export * from './focus';
 export * from './focus-visible';
 export * from './hover';
-export * from './interactions';
 export * from './press';
